api/pyproject.toml
[tool.poetry]
name = "parakeet-api"
version = "0.1.0"
description = "Parakeet — AI-powered incident response API"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
aiosqlite = "^0.20"
pydantic-settings = "^2.7"
google-adk = {extras = ["eval"], version = "^1.0"}
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
